+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.5.0
-* Add attribute popover support
+* Add attribute popover support. Closes #13.
+
+## 0.4.3
+* Fix bug for duplicated weekday label keys. Closes #28.
 
 ## 0.4.2
 * Convert event names to all lowercase to support in-DOM templates. Closes #26.